refactor: extract unit functionality outside of Metric component into a util
